Fix compute filters in unet (#313)
This PR fixes a bug in how filters are computed for the UNet backbone.
Previously, the calculation differed slightly from SLEAP due to rounding
errors, which caused a layer mismatch when running inference with legacy
SLEAP model weights. This update aligns the filter computation with
SLEAP’s implementation, ensuring compatibility with models trained in
the tensorflow framework.

Related Issue: #309

Co-authored-by: Divya Seshadri Murali <divyasesh@Mac-3475.lan>