Fix Inference on SuggestionFrames (#318)
This PR solves the following bug: When a user runs inference on
unlabeled suggested frame, `sleap-nn` raises an exception where
`self.labels.find(...)` has an empty list when filtering out unlabeled
suggested. The PR fixes the bug by initializing a new `LabeledFrame`
object with the given video and frame index and modify the condition to
check an empty list instead of indexing.