Add option to run tracking on select frames (#329)
This PR adds support for tracking on selected frames and specific video
index in the track-only pipeline. Previously, this was only possible
when tracking was run together with inference (to generate keypoint
predictions). With this PR, users can now specify frame indices or a
video index directly from an SLP file when running tracking
independently.