Auto-select GPUs (#333)
This PR adds functionality to automatically select GPUs with the most
available memory when the number of training devices specified by the
user is less than the total GPUs available. In this case, we query the
memory utilization of each GPU, sort them by free memory, and allocate
the devices with the highest availability.