Modify uv installation docs for cuda support (#340)
Previously, `uv tool install sleap-nn[torch-cuda128]` didn't install
torch with cuda support on windows (Related issue: #307). CUDA libraries
aren’t getting installed because PyPi doesn’t have access to
`tool.uv.index` . When we run `uv tool install` (installing `sleap-nn`
from PyPi), it defaults to pulling in the standard torch and torchvision
and for windows the default wheels is CPU (linux - CUDA if cuda is
available, mac - CPU). This happens because PyPi cannot see the custom
indices we provide in
[tool.uv.index](https://github.com/talmolab/sleap-nn/blob/99b80451f2af0758c2f944f7ebc0465b71b59c37/pyproject.toml#L108-L139)
(the metadata is not published with the wheel), so it cannot resolve
CUDA wheels.

`uv tool install` works when `--extra-index-url` is provided to point to
the appropriate Torch wheels. In this PR, we update the documentation to
include the required index URLs for installing with CUDA support.