Add warning for Python v3.14 (#348)
This PR adds warnings to the documentation that Python 3.14 is not yet
supported and updates install instructions to pin Python (e.g., --python
3.13) when using uv so users don’t accidentally end up on 3.14. It also
updates the CI workflows to run on Python 3.13. In this PR, we also
remove the deprecated flags `--extra-index-url` and `--index-url` for
`uv`-based installations and update to `--index` flag in the
installation docs.